Move certificate logic to separate file
